requirements.txt
streamlit
openai
streamlit-extras
python-dotenv 
plotly
scipy
scikit-learn
azure-search-documents==11.4.0b6